requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
aiogram==2.10.1
aiohttp==3.7.2
amqp==5.0.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.8.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.3.3
multidict==5.0.1
networkx==2.5
numpy==1.19.4
Pillow==8.0.1
prompt-toolkit==3.0.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
redis==3.5.3
requests==2.25.0
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.10.1
torch==1.4.0+cpu
torchvision==0.5.0+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.2
